fix: pair refacroting and claen up
