+++ README.md
@@ -1,7 +1,7 @@
 <br />
 <div align="center">
   <h1>Chargetrip examples</h1>
-  <p><h3 align="center">Ready-to-run example projects to get you started with Chargetrip API</h3></p>
+  <p><h3 align="center">Ready-to-run examples to get you started with Chargetrip API</h3></p>
   <a href="https://chargetrip.com">Website</a>
   <span>⚡️</span>
   <a href="https://docs.chargetrip.com/">Docs</a>
@@ -22,7 +22,7 @@ Available examples:
 2. Select a car [example](https://chargetrip.github.io/examples/car/);
 3. Show stations on a map [example](https://chargetrip.github.io/examples/stations/);
 4. Show all stations using the Vector Tile Server [example](https://chargetrip.github.io/examples/tile-server/);
-5. Build a route and display an Elevation plot [example](https://chargetrip.github.io/examples/route-path/)
+5. Build a route and display an elevation plot [example](https://chargetrip.github.io/examples/elevation-plot/).
 
 ## Installation and local development
