+++ elevation-plot/index.js
@@ -1,72 +1,7 @@
-import { createClient, createRequest, defaultExchanges, subscriptionExchange } from '@urql/core';
-import { pipe, subscribe } from 'wonka';
-import { SubscriptionClient } from 'subscriptions-transport-ws';
-import { createRoute, routeUpdate, getRoutePath } from './queries.js';
 import { drawRoute } from './map.js';
-import { loadGraph, imageLoader } from './graph';
 import * as mapboxPolyline from '@mapbox/polyline';
-import { displaySpecs, updateSpecs } from './journey-specs';
-
-/**
- * For the purpose of this example we use urgl - lightweights GraphQL client.
- * To establish a connection with Chargetrip GraphQL API you need to have an API key.
- * Read more about an authorisation in our documentation (https://docs.chargetrip.com/#authorisation).
- */
-const headers = {
-  'x-client-id': '5e8c22366f9c5f23ab0eff39',
-};
-
-const subscriptionClient = new SubscriptionClient('wss://api.chargetrip.io/graphql', {
-  reconnect: true,
-  connectionParams: headers,
-});
-
-const client = createClient({
-  url: 'https://api.chargetrip.io/graphql',
-  fetchOptions: {
-    method: 'POST',
-    headers,
-  },
-  exchanges: [
-    ...defaultExchanges,
-    subscriptionExchange({
-      forwardSubscription(operation) {
-        return subscriptionClient.request(operation);
-      },
-    }),
-  ],
-});
-
-/**
- * To create a route you need to:
- *
- * 1. create a new route and receive back its ID;
- * 2. subscribe to route updates in order to receive its details.
- */
-client
-  .mutation(createRoute)
-  .toPromise()
-  .then(response => {
-    const routeId = response.data.newRoute;
-
-    const { unsubscribe } = pipe(
-      client.executeSubscription(createRequest(routeUpdate, { id: routeId })),
-      subscribe(result => {
-        const { status, route } = result.data.routeUpdatedById;
-
-        // you can keep listening to the route changes to update route information
-        // for this example we want to only draw the initial route
-        if (status === 'done' && route) {
-          unsubscribe();
-          const routeData = result.data.routeUpdatedById.route;
-          drawRoutePolyline(routeData, response.data.newRoute);
-          loadGraph(routeData, routeData.elevationPlot);
-          imageLoader(routeData, routeData.legs);
-        }
-      }),
-    );
-  })
-  .catch(error => console.log(error));
+import { fetchRoute, fetchRoutePath } from './client';
+import { loadGraph, displayElevationData, displaySpecs } from './elevationGraph';
 
 /**
  * Draw a route on a map.
@@ -78,32 +13,22 @@ client
  * so we have to reverse every pair.
  *
  * @param data {object} route specification
+ * @param id {string} route ID
  */
+fetchRoute((routeId, routeData) => {
+  drawRoutePolyline(routeData, routeId);
+  loadGraph(routeData);
+});
+
 const drawRoutePolyline = (data, id) => {
   const decodedData = mapboxPolyline.decode(data.polyline);
   const reversed = decodedData.map(item => item.reverse());
+  const { elevationUp, elevationDown } = data;
 
-  client
-    .query(getRoutePath(id, reversed[0]))
-    .toPromise()
-    .then(response => {
-      displaySpecs(response.data);
-    })
-    .catch(error => console.log(error));
-  const map = drawRoute(reversed, data.legs);
-  updateSpecs(client, map, reversed, id);
-  displayRouteData(data);
-};
+  drawRoute(id, reversed, data.legs);
+  displayElevationData(elevationUp, elevationDown);
 
-/**
- * Show journey specific information like duration, consumption, costs etc.
- *
- * @param data {object} route specification
- */
-const displayRouteData = data => {
-  // The highest point in the elevation plot
-  document.getElementById('highest').innerHTML = `Up ${(data.elevationUp / 1000).toFixed(0)}km`;
-
-  // The lowest point in the elevation plot
-  document.getElementById('lowest').innerHTML = `Down ${(data.elevationDown / 1000).toFixed(0)}km`;
+  fetchRoutePath(id, reversed[0]).then(data => {
+    displaySpecs(data);
+  });
 };
