+++ public/elevation-plot/index.html
@@ -3,7 +3,7 @@
   <head>
     <title>Chargetrip API: Elevation plot example</title>
     <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-    <meta name="description" content="Build a route with the Chargetrip API" />
+    <meta name="description" content="Build a route with the Chargetrip API and show elevation data" />
     <meta name="keywords" content="Chargetrip, Chargetrip API, EV, ev-cars, ev-routing elevation consumption" />
     <meta name="viewport" content="initial-scale=1,maximum-scale=1,user-scalable=no" />
     <link href="https://fonts.googleapis.com/css2?family=Inter:wght@100;200;400;600&display=swap" rel="stylesheet" />
@@ -14,7 +14,7 @@
   <body>
     <div id="map"></div>
     <div class="github">
-      <a href="https://github.com/chargetrip/examples/route" title="github examples">
+      <a href="https://github.com/chargetrip/examples/elevation-plot" title="github examples">
         View on Github
       </a>
       <button aria-label="remove github link" title="remove github link" onclick="this.parentElement.remove()"></button>
@@ -27,16 +27,18 @@
         <h1>Journey Specs</h1>
         <div class="header-route">Amsterdam • Berlin</div>
         <div class="tags">
-          <div id="highest" class="tag"></div>
-          <div id="lowest" class="tag"></div>
+          <div id="highestPoint" class="tag"></div>
+          <div id="lowestPoint" class="tag"></div>
         </div>
         <div class="rounded-corners"></div>
       </header>
 
       <h2>Elevation levels</h2>
-      <div class="chart" id="chart"><canvas id="elevation" height="130px"></canvas></div>
-      <div class="line" id="line"></div>
-      <canvas class="charge" id="charge"></canvas>
+      <div class="chart">
+        <canvas id="elevationGraph" height="130px"></canvas>
+      </div>
+      <div class="line" id="elevationIndicator"></div>
+      <canvas class="charge" id="stationsLayer"></canvas>
       <h2>Details</h2>
 
       <ul class="info">
